ifs to switchcase and better converter

+++ cli/src/converter/items.js
@@ -2,7 +2,6 @@ const { existsSync, mkdirSync, writeFileSync, readFileSync } = require('fs')
 const jsonFormat = require('json-format')
 
 const properties = require('./properties')
-const naming = require('./naming')
 
 module.exports = doitems = (items, path) => {
   if (!items || !path) return
@@ -19,39 +18,36 @@ module.exports = doitems = (items, path) => {
 
     // if instance has no children
     if (!item.Item) {
-      if (!isFolder) {
-        if (existsSync(filepath(true))) {
-          if (isScript) {
-            writeFileSync(filepath(true), props.Source)
-
-            continue
-          }
-
-          let filedata = JSON.parse(
-            readFileSync(filepath(true), {
-              encoding: 'utf8',
-            })
-          )
-
-          if (typeof filedata === 'object') filedata = [filedata]
-
-          filedata.push(props)
-
-          writeFileSync(filepath(true), jsonFormat(filedata))
-
-          continue
-        }
-
-        writeFileSync(filepath(true), jsonFormat(props))
+      if (isScript) {
+        writeFileSync(filepath(true), props.Source)
 
         continue
       }
+
+      if (!isFolder && existsSync(filepath(true))) {
+        let filedata = JSON.parse(
+          readFileSync(filepath(true), {
+            encoding: 'utf8',
+          })
+        )
+
+        if (typeof filedata === 'object') filedata = [filedata]
+
+        filedata.push(props)
+
+        writeFileSync(filepath(true), jsonFormat(filedata))
+
+        continue
+      }
+
+      writeFileSync(filepath(true), jsonFormat(props))
 
       continue
     }
 
     if (!isFolder) {
       if (!existsSync(filepath())) mkdirSync(filepath())
+
       writeFileSync(
         `${filepath()}/index.${item['$'].class}`,
         isScript ? props.Source : jsonFormat(props)
